Move HtmlAttributesTrait methods into BasicWiget.
Given that all other widgets extends BasicWidget having the methods
in a trait was unnecessary.
